Change `count()` to type `long`.

+++ src/main/java/dev/root101/clean/core/app/usecase/CRUDUseCase.java
@@ -46,7 +46,7 @@ public interface CRUDUseCase<Domain extends DomainObject<ID>, ID> extends Abstra
      * @return findAll().size()
      * @throws RuntimeException
      */
-    public default int count() throws RuntimeException {
+    public default long count() throws RuntimeException {
         return findAll().size();
     }
